+++ Changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+-   The old API of `useHotKey()` was broken. Now the old and the new API are
+    supported. Using the old API just triggers a warning on the console.
+
 ## 6.4.1 - 2022-07-07
 
 ### Fixed